+++ CHANGELOG.md
@@ -17,6 +17,19 @@ This project adheres to [Semantic Versioning](http://semver.org/) and [Keep a Ch
 ### Breaks
 
 
+## Unreleased
+## 0.1.11 (2024-10-14)
+---
+
+### New
+Implement preamble model [GS-140].
+Implement ollama server [GS-139].
+Add AI_MODEL_ALLOW_SYSTEM_MSG, AI_MODEL_ALLOW_TOOLS, and AI_MODEL_NEED_PREAMBLE to manage models like Ollama "llava" that doesn't accept Tools [GS-140].
+
+### Changes
+Update ChatOllama adding the"langchain-ollama" dependency [GS-139].
+
+
 ## 0.1.10 (2024-10-07)
 ---